backend/requirements.txt
flask
flask-cors
requests
huggingface_hub
gunicorn
python-dotenv
